--- Pyllurium/Elements.py
@@ -1,281 +1,0 @@
-from Pyllurium.Element import Element
-
-
-class Hydrogen(Element):
-    Z = 1
-    symbol = 'H'
-
-
-class Helium(Element):
-    Z = 2
-    symbol = 'He'
-
-
-class Lithium(Element):
-    Z = 3
-    symbol = 'Li'
-
-
-class Beryllium(Element):
-    Z = 4
-    symbol = 'Be'
-
-
-class Boron(Element):
-    Z = 5
-    symbol = 'B'
-
-
-class Carbon(Element):
-    Z = 6
-    symbol = 'C'
-
-
-class Nitrogen(Element):
-    Z = 7
-    symbol = 'N'
-
-
-class Oxygen(Element):
-    Z = 8
-    symbol = 'O'
-
-
-class Flourine(Element):
-    Z = 9
-    symbol = 'F'
-
-
-class Neon(Element):
-    Z = 10
-    symbol = 'Ne'
-
-
-class Sodium(Element):
-    Z = 11
-    symbol = 'Na'
-
-
-class Magnesium(Element):
-    Z = 12
-    symbol = 'Mg'
-
-
-class Aluminum(Element):
-    Z = 13
-    symbol = 'Al'
-
-
-class Silicon(Element):
-    Z = 14
-    symbol = 'Si'
-
-
-class Phosphorous(Element):
-    Z = 15
-    symbol = 'P'
-
-
-class Sulfur(Element):
-    Z = 16
-    symbol = 'S'
-
-
-class Chlorine(Element):
-    Z = 17
-    symbol = 'Cl'
-
-
-class Argon(Element):
-    Z = 18
-    symbol = 'Ar'
-
-
-class Potassium(Element):
-    Z = 19
-    symbol = 'K'
-
-
-class Calcium(Element):
-    Z = 20
-    symbol = 'Ca'
-
-
-class Scandium(Element):
-    Z = 21
-    symbol = 'Sc'
-
-
-class Titanium(Element):
-    Z = 22
-    symbol = 'Ti'
-
-
-class Vanadium(Element):
-    Z = 23
-    symbol = 'V'
-
-
-class Chromium(Element):
-    Z = 24
-    symbol = 'Cr'
-
-
-class Manganese(Element):
-    Z = 25
-    symbol = 'Mn'
-
-
-class Iron(Element):
-    Z = 26
-    symbol = 'Fe'
-
-
-class Cobalt(Element):
-    Z = 27
-    symbol = 'Co'
-
-
-class Nickel(Element):
-    Z = 28
-    symbol = 'Ni'
-
-
-class Copper(Element):
-    Z = 29
-    symbol = 'Cu'
-
-
-class Zinc(Element):
-    Z = 30
-    symbol = 'Zn'
-
-
-class Gallium(Element):
-    Z = 31
-    symbol = 'Ga'
-
-
-class Germanium(Element):
-    Z = 32
-    symbol = 'Ge'
-
-
-class Arsenic(Element):
-    Z = 33
-    symbol = 'As'
-
-
-class Selenium(Element):
-    Z = 34
-    symbol = 'Se'
-
-
-class Bromine(Element):
-    Z = 35
-    symbol = 'Br'
-
-
-class Krypton(Element):
-    Z = 36
-    symbol = 'Kr'
-
-
-class Rubidium(Element):
-    Z = 37
-    symbol = 'Rb'
-
-
-class Strontium(Element):
-    Z = 38
-    symbol = 'Sr'
-
-
-class Yttrium(Element):
-    Z = 39
-    symbol = 'Y'
-
-
-class Zirconium(Element):
-    Z = 40
-    symbol = 'Zr'
-
-
-class Niobium(Element):
-    Z = 41
-    symbol = 'Nb'
-
-
-class Molybdenum(Element):
-    Z = 42
-    symbol = 'Mo'
-
-
-class Technetium(Element):
-    Z = 43
-    symbol = 'Tc'
-
-
-class Ruthenium(Element):
-    Z = 44
-    symbol = 'Ru'
-
-
-class Rhodium(Element):
-    Z = 45
-    symbol = 'Rh'
-
-
-class Palladium(Element):
-    Z = 46
-    symbol = 'Pd'
-
-
-class Silver(Element):
-    Z = 47
-    symbol = 'Ag'
-
-
-class Cadmium(Element):
-    Z = 48
-    symbol = 'Cd'
-
-
-class Indium(Element):
-    Z = 49
-    symbol = 'In'
-
-
-class Tin(Element):
-    Z = 50
-    symbol = 'Sn'
-
-
-class Antimony(Element):
-    Z = 51
-    symbol = 'Sb'
-
-
-class Tellurium(Element):
-    Z = 52
-    symbol = 'Te'
-
-
-class Iodine(Element):
-    Z = 53
-    symbol = 'I'
-
-
-class Xenon(Element):
-    Z = 54
-    symbol = 'Xe'
-
-
-class Cesium(Element):
-    Z = 55
-    symbol = 'Cs'
-
-
-class Barium(Element):
-    Z = 56
-    symbol = 'Ba'
